requirements.txt
langchain==0.0.188
python-dotenv==1.0.0
streamlit==1.23.1
openai==0.27.7
faiss-cpu==1.7.4
pytube==15.0.0